Refactored GF Embedded tests - set work dir inside target directory for creating temporary files - renamed tests that end in ITest to IUnitTest to run them as UT not IT Signed-off-by:Ondro Mihalyi <mihalyi@omnifish.ee>
