+++ appserver/tests/embedded/glassfish_resources_xml/src/test/java/org/glassfish/tests/embedded/cdi_ejb_jpa/BasicCDIUnitTest.java
@@ -1,5 +1,4 @@
 /*
- * Copyright (c) 2023 Contributors to the Eclipse Foundation.
  * Copyright (c) 2011, 2018 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
@@ -15,11 +14,13 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package org.glassfish.tests.embedded.cdi_basic;
+package org.glassfish.tests.embedded.cdi_ejb_jpa;
 
+import org.junit.jupiter.api.Assertions;
 import org.glassfish.embeddable.Deployer;
 import org.glassfish.embeddable.GlassFish;
 import org.glassfish.embeddable.GlassFishProperties;
+import org.glassfish.embeddable.BootstrapProperties;
 import org.glassfish.embeddable.GlassFishRuntime;
 import org.glassfish.embeddable.archive.ScatteredArchive;
 import org.glassfish.embeddable.web.HttpListener;
@@ -37,27 +38,29 @@ import java.util.Date;
 import java.util.Enumeration;
 import java.util.jar.JarEntry;
 import java.util.jar.JarFile;
-import org.junit.jupiter.api.Assertions;
 
 /**
  * @author bhavanishankar@dev.java.net
  */
 
-public class BasicCDITest{
+public class BasicCDIUnitTest {
+
+    private static final String PROJECT_DIR = System.getProperty("project.directory");
 
     @Test
     public void test() throws Exception {
 
         GlassFishProperties props = new GlassFishProperties();
+        BootstrapProperties bootstrapProperties = new BootstrapProperties();
         props.setPort("http-listener", 8080);
-        GlassFish glassfish = GlassFishRuntime.bootstrap().newGlassFish(props);
+        GlassFish glassfish = GlassFishRuntime.bootstrap(bootstrapProperties).newGlassFish(props);
         glassfish.start();
 
         // Test Scattered Web Archive
-        ScatteredArchive sa = new ScatteredArchive("cdi_basic",
-                ScatteredArchive.Type.WAR, new File("src/main/webapp"));
-        sa.addClassPath(new File("target/classes"));
-        sa.addClassPath(new File("src/main/resources"));
+        ScatteredArchive sa = new ScatteredArchive("cdi_ejb_jpa",
+                ScatteredArchive.Type.WAR, new File(PROJECT_DIR, "src/main/webapp"));
+        sa.addClassPath(new File(PROJECT_DIR, "target/classes"));
+        sa.addClassPath(new File(PROJECT_DIR, "src/main/resources"));
         URI warURI = sa.toURI();
         printContents(warURI);
 
@@ -65,7 +68,7 @@ public class BasicCDITest{
         Deployer deployer = glassfish.getDeployer();
         String appname = deployer.deploy(warURI);
         System.out.println("Deployed [" + appname + "]");
-        Assertions.assertEquals(appname, "cdi_basic");
+        Assertions.assertEquals(appname, "cdi_ejb_jpa");
 
         // Now create a http listener and access the app.
         WebContainer webcontainer = glassfish.getService(WebContainer.class);
@@ -74,7 +77,7 @@ public class BasicCDITest{
         listener.setPort(9090);
         webcontainer.addWebListener(listener);
 
-        get("http://localhost:8080/cdi_basic/BasicCDITestServlet",
+        get("http://localhost:8080/cdi_ejb_jpa/BasicCDITestServlet",
                 "All CDI beans have been injected.");
 
         deployer.undeploy(appname);
